Fixing bug related to the values type

+++ penrose/funcsLine.py
@@ -26,5 +26,5 @@ def basic_incremental(vertices):
 
 
 def bresenham(vertices):
-	return np.array(list(bh_algorithm(*tuple(np.abs(vertices).reshape(1, 4)[0]))))
+	return np.array(list(bh_algorithm(*tuple(np.uint8(np.abs(vertices)).reshape(1, 4)[0]))))
